chore: remove lodash export
